# src/main/java/com/fiec/estoqueback/features/user/dto/CreatedUserResponseDto.java
package com.fiec.estoqueback.features.user.dto;

import lombok.AllArgsConstructor;
import lombok.Builder;
import lombok.Data;
import lombok.NoArgsConstructor;

@Data
@Builder
@NoArgsConstructor
@AllArgsConstructor
public class CreatedUserResponseDto {
    String id;
    String userId;
}